#33 Rename from literacyapp to elimu
- "org.literacyapp" --> "ai.elimu"
