+++ README.md
@@ -1,4 +1,4 @@
-# literacyapp-start-guide
+# start-guide
 Android application teaching basic tablet UI navigation
 
 ![screenshot_20170605-194552](https://cloud.githubusercontent.com/assets/15718174/26795982/fee64200-4a27-11e7-9c2a-e000efefda3a.png)
