+++ app/src/main/java/ai/elimu/startguide/receiver/BootReceiver.java
@@ -1,0 +1,18 @@
+package ai.elimu.startguide.receiver;
+
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+
+public class BootReceiver extends BroadcastReceiver {
+
+    @Override
+    public void onReceive(Context context, Intent intent) {
+//        if (Intent.ACTION_BOOT_COMPLETED.equals(intent.getAction()) && StartPrefsHelper.startAfterBoot(context)) {
+//            Intent bootIntent = new Intent(context, SwipeUpDownActivity.class);
+//            bootIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+//            context.startActivity(bootIntent);
+//        }
+    }
+}
