+++ app/src/main/res/layout/activity_exit_full_screen.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/light_blue"
-    tools:context="org.literacyapp.startguide.content.ExitFullScreenActivity">
+    tools:context="ai.elimu.startguide.content.ExitFullScreenActivity">
 
     <ImageView
         android:layout_width="match_parent"
@@ -14,7 +14,7 @@
         android:src="@drawable/landscape"
         android:scaleType="centerCrop"/>
 
-    <org.literacyapp.handgesture.HandView
+    <ai.elimu.handgesture.HandView
         android:id="@+id/hand_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -53,7 +53,7 @@
             android:src="@drawable/ic_recents"/>
     </RelativeLayout>
 
-    <org.literacyapp.handgesture.HandView
+    <ai.elimu.handgesture.HandView
         android:id="@+id/hand_view_bottom"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
