+++ app/src/main/res/layout/activity_final.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/white"
-    tools:context="org.literacyapp.startguide.content.FinalActivity">
+    tools:context="ai.elimu.startguide.content.FinalActivity">
 
     <ImageView
         android:id="@+id/final_checkmark"
