+++ app/src/main/res/layout/activity_swipe_up_down.xml
@@ -7,7 +7,7 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    tools:context="org.literacyapp.startguide.content.swipe.SwipeUpDownActivity" >
+    tools:context="ai.elimu.startguide.content.swipe.SwipeUpDownActivity" >
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerView"
@@ -15,7 +15,7 @@
         android:layout_height="wrap_content"
         tools:listitem="@layout/item_main"/>
 
-    <org.literacyapp.handgesture.HandView
+    <ai.elimu.handgesture.HandView
         android:id="@+id/hand"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
