+++ app/src/main/res/layout/activity_test_gesture.xml
@@ -5,29 +5,29 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.literacyapp.startguide.TestGestureActivity" >
+    tools:context="ai.elimu.startguide.TestGestureActivity" >
 
-    <org.literacyapp.handgesture.HandView
+    <ai.elimu.handgesture.HandView
         android:id="@+id/single_tap"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_margin="30dp" />
 
-    <org.literacyapp.handgesture.HandView
+    <ai.elimu.handgesture.HandView
         android:id="@+id/double_tap"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_margin="30dp"
         android:layout_alignParentRight="true"/>
 
-    <org.literacyapp.handgesture.HandView
+    <ai.elimu.handgesture.HandView
         android:id="@+id/press_and_hold"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_margin="30dp"
         android:layout_below="@id/single_tap" />
 
-    <org.literacyapp.handgesture.HandView
+    <ai.elimu.handgesture.HandView
         android:id="@+id/translation"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
