+++ handgesture/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="org.literacyapp.handgesture">
+          package="ai.elimu.handgesture">
 
     <application android:allowBackup="true"
                  android:label="@string/app_name"
