+++ handgesture/src/main/java/ai/elimu/handgesture/Constants.java
@@ -1,0 +1,14 @@
+package ai.elimu.handgesture;
+
+/**
+ */
+public class Constants {
+
+    public static final int MILLISECONDS = 1000;
+
+    public static final long DEFAULT_ANIMATION_DELAY = 1; //seconds
+
+    public static final long SCALE_DURATION = 500; //milliseconds
+
+    public static final float SCALE_FACTOR = 0.9f;
+}
