+++ handgesture/src/main/java/ai/elimu/handgesture/HandGestureListener.java
@@ -1,0 +1,14 @@
+package ai.elimu.handgesture;
+
+/**
+ */
+public interface HandGestureListener {
+
+    void onAnimationStarted();
+
+    void onZoomOutEnd();
+
+    void onTouchStart();
+
+    void onTouchEnd();
+}
