+++ handgesture/src/main/java/ai/elimu/handgesture/HandViewListener.java
@@ -1,0 +1,6 @@
+package ai.elimu.handgesture;
+
+public interface HandViewListener {
+
+    void onHandAnimationEnd();
+}
