--- handgesture/src/main/java/org/literacyapp/handgesture/HandGestureListener.java
@@ -1,14 +1,0 @@
-package org.literacyapp.handgesture;
-
-/**
- */
-public interface HandGestureListener {
-
-    void onAnimationStarted();
-
-    void onZoomOutEnd();
-
-    void onTouchStart();
-
-    void onTouchEnd();
-}
