--- handgesture/src/main/java/org/literacyapp/handgesture/HandViewListener.java
@@ -1,6 +1,0 @@
-package org.literacyapp.handgesture;
-
-public interface HandViewListener {
-
-    void onHandAnimationEnd();
-}
